fix(all): all packages now output both .mjs and .js and use them in their exports (#669)
In this PR we offer both a `.mjs` as well as a js build, this so that we
can put mjs in the export-maps so we can adhere to strict node
resolution for all node-versions which need mjs to show them that we are
using import/export. The reason for also supplying a js here is because
tamagui uses dist/esm in the module property. This will make it so that
older webpack/.. versions can still automatically pick what they can
resolve according to their configuration.